fix: initializEable -> initializAble
